Final name and lib structure, display version
and release date in OpenProject footer

- used translations from old customize plugin
- changed source path of parsed files to rails_root/config
