+++ README.rdoc
@@ -1,3 +1,3 @@
-= OpenprojectRelease
+= OpenProject::Release
 
 This project rocks and uses MIT-LICENSE.
