+++ config/locales/de.yml
@@ -1,0 +1,3 @@
+de:
+  op_release_last_update: "Letzte Aktualisierung %{released_at}"
+  op_release_current_version: "Version %{version}"
