+++ lib/open_project/release/version.rb
@@ -1,0 +1,5 @@
+module OpenProject
+  module Release
+    VERSION = "0.2.5"
+  end
+end
