+++ lib/openproject-release.rb
@@ -1,0 +1,1 @@
+require 'open_project/release'
