--- lib/openproject_release/version.rb
@@ -1,3 +1,0 @@
-module OpenprojectRelease
-  VERSION = "0.2.5"
-end
